Rename gluon kernels and remove unused versions
